Save and continue when there is no feedback
Why do it this way?

`onAnswerSave` is non-blocking and non-awaitable. Consequently, calling
`onAnswerSave` in the onClick of the submit button does not work
because this will advance to the next question before the current one
is finished saving.

We know that the api is done saving when `is_completed` becomes true.

When `hasFeedback` is falsy and the submit button is clicked,
`shouldContinue` is set to `true`.

If both `shouldContinue` and `is_completed` are true, then the answer
has been submitted and saved and it is okay to advance.

[CORE-173]